+++ ambition_utils/docs/release_notes.rst
@@ -1,6 +1,14 @@
 Release Notes
 =============
 
+2.5.3
+-----
+* Fixed cloning when until param is present
+
+2.5.2
+-----
+* Fixed cloning with bynweekday data
+
 2.5.1
 -----
 * Updated the get_time_zone() helper method with pytz.all_timezones